requirements.txt
dash 
dash-renderer 
dash-html-components 
dash-core-components 
plotly
pandas-datareader
praw==3.6.0
redis
mysqlclient